added haptic feedback, models separated to files, added buttonBlock function
